% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addCountsOfFeatureClasses}
\alias{addCountsOfFeatureClasses}
\title{Count occurrences of feature classes}
\usage{
addCountsOfFeatureClasses(summary, feature_classes)
}
\arguments{
\item{summary}{Summary table generated by \link[DEUS]{mergeResults} (requires BLAST result generated by \link[DEUS]{runBlast})}

\item{feature_classes}{List of features representing classes to be counted.
Features can be defined as regular expressions, as described in \link[stringi]{stringi-search-regex}.}
}
\value{
The summary table with additional columns for each feature term given by the user.
Each column holds the number of features in 'FeatureList' that match the corresponding feature term.
The column 'Other' reflects the number of all features in the 'FeatureList' after substracting the sum of feature class matches.
}
\description{
As each unique sequence can have a lot of BLAST hits this function counts the number of hits grouped by user defined feature classes.
}
