% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counting.R
\name{createCountTableFromFastQs}
\alias{createCountTableFromFastQs}
\title{Create sequence-based count tables}
\usage{
createCountTableFromFastQs(in_dir, fq_suffix = ".fq.gz", pheno_info = NULL)
}
\arguments{
\item{in_dir}{Input directory containing fastq files to be analyzed}

\item{fq_suffix}{Suffix of your fastq files, default is ".fq.gz"}

\item{pheno_info}{A data frame with sample names in column 'sampleName' and file names in column 'sample'. The sample names will be used as headers in the count table. If not given count table headers will be the file names of the input fastQ files.}
}
\description{
This function creates count tables summarizing counts for each observed sequence in a set of fastQ files.
Input files are taken from the given input directory and are selected based on the given fastQ suffix.
}
