% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createMap}
\alias{createMap}
\title{Create a mapping of sequences to sequence identifiers}
\usage{
createMap(count_table)
}
\arguments{
\item{count_table}{The count table generated by \link[DEUS]{createCountTableFromFastQs}}
}
\value{
A data frame with sequences as row names and sequence identifiers in the column 'SequenceID'
}
\description{
This function creates a data frame mapping the row names of the input (sequence) count table to sequence identifiers.
In context of the DEUS package the sequence identifiers can be used as alternate short identifiers for each unique sequence.
}
