% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deleteTmp}
\alias{deleteTmp}
\title{Remove temporary FASTA file required for clustering}
\usage{
deleteTmp(out_dir)
}
\arguments{
\item{out_dir}{Directory where output files of clustering are saved}
}
\value{
Deletes the file sig_sequences.fa
}
\description{
This function is internally called by \link[DEUS]{runClustering}
}
