% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{generateSummaryPlots}
\alias{generateSummaryPlots}
\title{TODO Function to generate plots for SummaryTable}
\usage{
generateSummaryPlots(summary, classes, out_dir)
}
\arguments{
\item{summary}{summary table generated by mergeResults function (requires BLAST result)}

\item{classes}{list of prefixes of features representing classes to be included in the plots}

\item{out_dir}{directory for plots (boxplot and another one for each class)}
}
\description{
This function generates a boxplot showing the distribution of BLAST hits for each sncRNA class of interest.
Further it plots the number of BLAST hits as a function of the length of the sequences and performs linear regression.
}
\keyword{statistics}
\keyword{summary}
