% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generateSummaryStats}
\alias{generateSummaryStats}
\title{Function to generate stats on summary}
\usage{
generateSummaryStats(summary, classes)
}
\description{
Function to generate stats on summary
}
\keyword{statistics}
\keyword{summary}
