% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getConditionCountStats}
\alias{getConditionCountStats}
\title{Function to compute mean and sd of normalized counts for each condition}
\usage{
getConditionCountStats(countData, phenoData)
}
\arguments{
\item{countData}{table of (normalized) counts per sequence}

\item{phenoData}{data frame with sample names as rownames and assigned condition in first column}
}
\description{
Function to compute mean and sd of normalized counts for each condition
}
\keyword{count}
\keyword{statistics}
