% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getNoBlastHitFraction}
\alias{getNoBlastHitFraction}
\title{Function to generate plots for SummaryTable}
\usage{
getNoBlastHitFraction(summary, countTable = NULL, phenoInfo = NULL)
}
\description{
Function to generate plots for SummaryTable
}
\keyword{statistics}
\keyword{summary}
