% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mergeAndAggregate}
\alias{mergeAndAggregate}
\title{Aggregates sequence counts for each sequence cluster}
\usage{
mergeAndAggregate(map, cl_count_table, clustResult)
}
\arguments{
\item{map}{A data frame with sequences as row names and sequence identifiers in first column.
Can be generated by \link[DEUS]{createMap}}

\item{cl_count_table}{Count table created by \link[DEUS]{createCountTableFromFastQs}}

\item{clustResult}{Dataframe containing clustering results as created by \link[DEUS]{runClustering}}
}
\value{
A dataframe that includes sequence counts for each sequence cluster. Can be used as input for DE analysis
}
\description{
Aggregates sequence counts for each sequence cluster
}
