% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{plotPCA}
\alias{plotPCA}
\title{Function to plot PCA}
\usage{
plotPCA(rld, pheno_info, out_dir)
}
\arguments{
\item{rld}{regularized log transformed counts}

\item{pheno_info}{data frame with sample names as rownames and assigned condition in first column}

\item{out_dir}{directory for sample distance map}
}
\description{
Function to plot PCA
}
\keyword{distance}
\keyword{map}
\keyword{sample}
