% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{plotSampleDistanceMap}
\alias{plotSampleDistanceMap}
\title{Function to plot sample distance map}
\usage{
plotSampleDistanceMap(rld, out_dir)
}
\arguments{
\item{rld}{regularized log transformed counts}

\item{out_dir}{directory for sample distance map}
}
\description{
Function to plot sample distance map
}
\keyword{distance}
\keyword{map}
\keyword{sample}
