% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{printCluster}
\alias{printCluster}
\title{Print FASTA file for cluster of sequences}
\usage{
printCluster(data, cl_dir)
}
\arguments{
\item{data}{Subset of overall data provided within the \link[DEUS]{processClusters} function}

\item{cl_dir}{Directory of CD-HIT result file and where generated files will be saved}
}
\value{
A FASTA file in the Clusters folder in cl_dir representing the sequences of a cluster.
}
\description{
This function is internally called by \link[DEUS]{processClusters}
}
