% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{printClusterSummary}
\alias{printClusterSummary}
\title{Prints a summary of the number of significant sequences and clusters}
\usage{
printClusterSummary(summary, sig_threshold = 0.05)
}
\arguments{
\item{summary}{Summary table as created by \link[DEUS]{mergeResults}}

\item{sig_treshold}{Significance threshold used for filtering according to the IHW P-value of the sequences and clusters}
}
\value{
Prints summary values to standard output.
}
\description{
Prints a summary of the number of significant sequences and clusters
}
