% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{processClusters}
\alias{processClusters}
\title{TODO Processes CD-Hit Outfile}
\usage{
processClusters(map, clusters, out_dir)
}
\arguments{
\item{map}{data frame with sequences as row names and some identifier for each sequence in the first column}

\item{out_dir}{Output folder}
}
\description{
TODO Processes CD-Hit Outfile
}
\keyword{clustering}
