% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{runClustering}
\alias{runClustering}
\title{Function to execute CD-HIT Clustering}
\usage{
runClustering(cdhit_path, sequences, out_dir, identityCutoff, lengthCutoff,
  wordlength, optional = "")
}
\arguments{
\item{cdhit_path}{Path to cd-hit-est executable}

\item{sequences}{File including the sequences in FASTA format}

\item{out_dir}{Output folder}

\item{identityCutoff}{Sequence identity cutoff used for clustering}

\item{lengthCutoff}{Length difference cutoff}

\item{wordlength}{CD-Hit word length}

\item{optional}{Optional execution parameters}
}
\description{
Function to execute CD-HIT Clustering
}
\keyword{clustering}
