% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sequencesAsFasta}
\alias{sequencesAsFasta}
\title{Create a vector of sequence identifiers and sequences in FASTA format}
\usage{
sequencesAsFasta(de_result, map)
}
\arguments{
\item{de_result}{A data frame with sequences as row names.
In the context of DEUS, the data frame is the resulting table of differential expression analysis with significant sequences as row names.
Differential expression analysis can be performed via \link[DEUS]{runDESeq2}}

\item{map}{A data frame with sequences as row names and sequence identifiers in first column.
Can be generated by \link[DEUS]{createMap}}
}
\value{
A vector of alternating sequence identifiers and nucleotide sequences
}
\description{
Create a vector of sequence identifiers and sequences in FASTA format
}
