% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{writeSummaryFiles}
\alias{writeSummaryFiles}
\title{Write summary tables for DEUS result}
\usage{
writeSummaryFiles(summary, out_dir, expressed_only = TRUE)
}
\arguments{
\item{summary}{Summary table generated by \link[DEUS]{mergeResults} (requires BLAST result generated by \link[DEUS]{runBlast})}

\item{out_dir}{Directory where summary tables are written}

\item{expressed_only}{If true (default), only sequences with an IHWPvalue are provided in the output file. Sequences without IHWPvalue were not part of the DE analysis and therefore are assumed to have low expression values. If filtering is turned off, output files might include a large number of rows.}
}
\value{
Files are written to the output directory. 'SummaryTable.tsv' corresponds to the table generated by \link[DEUS]{mergeResults}.
The table is split into a file 'SummaryTable_withBlast.tsv' and a file 'SummaryTable_noBlast.tsv' representing the sequences that have a BLAST hit and those with no BLAST hit.
}
\description{
Write summary tables for DEUS result
}
