% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addCountsOfFeatureClasses}
\alias{addCountsOfFeatureClasses}
\title{Function to add counts of different feature classes}
\usage{
addCountsOfFeatureClasses(mergedResult, featureClasses)
}
\arguments{
\item{mergedResult}{summary table including blast results}

\item{featureClasses}{List of features representing classes to be counted.
Features can be defined as regular expressions, as described in \link[stringi]{stringi-search-regex}.}
}
\description{
Function to add counts of different feature classes
}
\keyword{counting}
\keyword{feature}
