% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addCountsOfFeatureClasses}
\alias{addCountsOfFeatureClasses}
\title{Function add counts of different feature classes}
\usage{
addCountsOfFeatureClasses(mergedResult, featureClasses)
}
\description{
Function add counts of different feature classes
}
\examples{

}
\keyword{counting}
\keyword{feature}
