% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createMap}
\alias{createMap}
\title{Function to create a map for row IDs of a data frame}
\usage{
createMap(countTable)
}
\arguments{
\item{countTable}{the table generated by createCountTableFromFastQs}
}
\description{
This function creates a map for row IDs of its input data frame.
In context of the USBseq package the sequence IDs can be used as alternate short identifiers for each unique sequence.
}
\examples{
in_dir <- system.file("extdata", package = "USBseq")
countTable <- createCountTableFromFastQs(in_dir)
map <- createMap(countTable)
}
\keyword{IDs}
\keyword{sequence}
