% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deleteTmp}
\alias{deleteTmp}
\title{Function to remove temporary files after pipeline execution}
\usage{
deleteTmp(outDir)
}
\arguments{
\item{outDir}{output folder}
}
\description{
Function to remove temporary files after pipeline execution
}
\examples{

}
\keyword{clean-up}
