% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filterLowExp}
\alias{filterLowExp}
\title{Function to filter a count table}
\usage{
filterLowExp(countData, pheno)
}
\arguments{
\item{countTable}{the table generated by createCountTableFromFastQs}

\item{phenofile}{data frame with sample IDs as row.names and column 'condition'. Sample IDs must be identical to those in countTable.}
}
\description{
This function pre-filters a count table of sequences using additional information on sample conditions.
Filtering is especially useful in context of the DEUS package because sequence based count tables are generally very large.
Sequences are kept if both groups have an average sequence count of at least 10 or if one of both group has zero counts for all samples.
}
\examples{
in_dir <- system.file("extdata", package = "DEUS")
countTable <- createCountTableFromFastQs(in_dir)
phenoInfo <- read.table(system.file("extdata", "condition.tsv", package = "DEUS"), header=T, row.names=1)
countTableFiltered <- filterLowExp(countTable, phenoInfo)
}
\keyword{filtering}
