% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filterLowExp}
\alias{filterLowExp}
\title{Function to filter a count table}
\usage{
filterLowExp(countData, phenofile)
}
\arguments{
\item{phenofile}{path to tab separated file with first column 'sample' and second column 'condition'. Sample IDs must be identical to those in countTable.}

\item{countTable}{the table generated by createCountTableFromFastQs}
}
\description{
This function pre-filters a count table of sequences using additional information on sample conditions.
Filtering is especially useful in context of the USBseq package because sequence based count tables are generally very large.
Sequences are kept if both groups have an average sequence count of at least 10 or if one of both group has zero counts for all samples.
}
\examples{
in_dir <- system.file("extdata", package = "USBseq")
countTable <- createCountTableFromFastQs(in_dir)
phenofile <- "./my_experiment/condition.tsv"
countTableFiltered <- filterLowExp(countTable, phenofile)
}
\keyword{filtering}
