% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counting.R
\name{filterLowExp}
\alias{filterLowExp}
\title{Filter a count table}
\usage{
filterLowExp(count_table, pheno_info, group_threshold = 10)
}
\arguments{
\item{count_table}{The count table generated by \link[DEUS]{createCountTableFromFastQs}}

\item{pheno_info}{A data frame with sample identifiers as row names including the column 'condition'.
The sample identifiers must be identical to those in the count_table.}

\item{group_threshold}{Minimum average sequence count for both groups}
}
\value{
The input count table with all sequences removed that fail the filter critera
}
\description{
This function filters a table of (sequence) counts using additional information on sample conditions.
Filtering is especially useful in context of the DEUS package because sequence based count tables are generally very large.
Sequences are kept if both groups have a default minimum average sequence count of 10 or if one of both groups has zero counts for all samples.
}
