% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generateSummaryPlots}
\alias{generateSummaryPlots}
\title{Function to generate plots for SummaryTable}
\usage{
generateSummaryPlots(summary, classes, out_dir)
}
\description{
Function to generate plots for SummaryTable
}
\keyword{statistics}
\keyword{summary}
