% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getConditionCountStats}
\alias{getConditionCountStats}
\title{Compute mean and standard deviation of normalized counts for each condition}
\usage{
getConditionCountStats(count_table, pheno_info,
  condition_col = "condition")
}
\arguments{
\item{count_table}{A table of (normalized) sequence counts.
A sequence count table can be generated by \link[DEUS]{createCountTableFromFastQs} and normalized by applying \link[DEUS]{runDESeq2}}

\item{pheno_info}{A data frame with sample names as row names and one or more condition columns.}

\item{condition_col}{Name of the column that should be used as the condition to group the samples.}
}
\value{
A data frame with sequences as row names and the mean and standard deviation for each condition defined in pheno_info
}
\description{
Compute mean and standard deviation of normalized counts for each condition
}
