% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{getNoBlastHitFraction}
\alias{getNoBlastHitFraction}
\title{Compute fraction of reads without BLAST hit for each sample}
\usage{
getNoBlastHitFraction(summary, count_table)
}
\arguments{
\item{summary}{Summary table generated by \link[DEUS]{mergeResults} (requires BLAST result generated by \link[DEUS]{runBlast})}

\item{count_table}{A table of (normalized) sequence counts.
A sequence count table can be generated by \link[DEUS]{createCountTableFromFastQs} and normalized by applying \link[DEUS]{runDESeq2}}
}
\value{
A data frame with samples as row names and the column 'NA_fraction'
}
\description{
Compute fraction of reads without BLAST hit for each sample
}
