% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{getNoBlastHitFraction}
\alias{getNoBlastHitFraction}
\title{TODO Function to compute fraction of reads without BLAST hit for each sample}
\usage{
getNoBlastHitFraction(summary, countTable)
}
\arguments{
\item{summary}{summary table generated by mergeResults function (requires BLAST result)}

\item{countTable}{table of (normalized) counts per sequence per sample}
}
\description{
TODO Function to compute fraction of reads without BLAST hit for each sample
}
\keyword{statistics}
\keyword{summary}
