% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mergeResults}
\alias{mergeResults}
\title{TODO Function to merge DEA, blast result}
\usage{
mergeResults(deResult = NULL, countStats = NULL, blastResult = NULL,
  clustResult = NULL, map)
}
\arguments{
\item{deResult}{result of differential expression analysis (columns log2FoldChange, pvalue, IHWPval are required)}

\item{countStats}{result of getConditionCountStats function}

\item{blastResult}{result of blast function}

\item{clustResult}{result of clustering function}

\item{map}{data frame with sequences as row names and sequence IDs in first column}
}
\description{
TODO Function to merge DEA, blast result
}
\keyword{generating}
\keyword{summary}
\keyword{tables}
