% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mergeResults}
\alias{mergeResults}
\title{Merge intermediate results of DEUS}
\usage{
mergeResults(de_result = NULL, count_stats = NULL,
  blast_result = NULL, clust_result = NULL, map)
}
\arguments{
\item{de_result}{Result of differential expression analysis generated by \link[DEUS]{runDESeq2} (columns 'log2FoldChange', 'pvalue' and 'IHWPval' are required)}

\item{count_stats}{Means and standard deviations of each condition in analysis generated by \link[DEUS]{getConditionCountStats}}

\item{blast_result}{Result of BLAST generated by \link[DEUS]{runBlast} (columns 'qseqid', 'sseqid', 'length' and 'evalue' are required)}

\item{clust_result}{Result of clustering generated by \link[DEUS]{runClustering}}

\item{map}{A data frame with sequences as row names and sequence identifiers in first column.
Can be generated by \link[DEUS]{createMap}}
}
\value{
If all results are given, a data frame is created with sequences as row names and the columns 'SequenceID', 'Log2FoldChange', 'Pvalue', 'IHWPvalue', 'NormCounts_<cond1>_Mean', 'NormCounts_<cond1>_Sd', 'NormCounts_<cond2>_Mean', 'NormCounts_<cond2>_Sd', 'ClusterID', 'Length', 'BlastEvalue' and 'FeatureList'.
If any result is not given corresponding columns are missing.
}
\description{
The function merges results created during DEUS analysis into a comprehensive summary table.
It is also possible to merge only a sub-set of all results or to generate the results with other methods than the provided ones.
}
