% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mergeSingleAndClusterResults}
\alias{mergeSingleAndClusterResults}
\title{Merges clustering results as well as DE results for single sequence and cluster approach}
\usage{
mergeSingleAndClusterResults(cl_sig_results, clust_result, sig_results,
  map)
}
\arguments{
\item{cl_sig_results}{A data frame as created by \link[DEUS]{runDESeq2}. Includes DE pvalues for each cluster}

\item{clust_result}{Dataframe containing clustering results as created by \link[DEUS]{runClustering}}

\item{sig_results}{A data frame as created by \link[DEUS]{runDESeq2}. Includes DE pvalues for each individual sequence}

\item{map}{A data frame with sequences as row names and sequence identifiers in first column.
Can be generated by \link[DEUS]{createMap}}
}
\value{
A combined data frame which includes all pvalues and clusterIDs
}
\description{
Merges clustering results as well as DE results for single sequence and cluster approach
}
