% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{plotPCA}
\alias{plotPCA}
\title{Plot result of PCA}
\usage{
plotPCA(rld, pheno_info, out_dir, prefix = "DESeq2")
}
\arguments{
\item{rld}{Count table after 'regularized log transformation' by \link[DESeq2]{rlog}}

\item{pheno_info}{A data frame with sample identifiers as row names including maximal two columns defining (sub-)conditions.
The sample identifiers must be identical to those in the rld table.}

\item{out_dir}{Directory for PDF file including the PCA plot}

\item{prefix}{Prefix for name of PDF file}
}
\value{
DESeq2_PCA.pdf in output directory
}
\description{
This function is internally called by \link[DEUS]{runDESeq2}
}
