% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{plotSampleDistanceMap}
\alias{plotSampleDistanceMap}
\title{Plot sample distance heatmap}
\usage{
plotSampleDistanceMap(rld, out_dir)
}
\arguments{
\item{rld}{Count table after 'regularized log transformation' by \link[DESeq2]{rlog}}

\item{out_dir}{Directory for PDF file including the sample distance heatmap}
}
\value{
DESeq2_sample_dist.pdf in output directory
}
\description{
This function is internally called by \link[DEUS]{runDESeq2}
}
