% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{printCluster}
\alias{printCluster}
\title{TODO Helper function for generating individual cluster fastas, not meant for external use}
\usage{
printCluster(data, cl.dir)
}
\arguments{
\item{data}{Subset of overall data provided within the processClusters function}

\item{cl.dir}{Cluster output folder}
}
\description{
TODO Helper function for generating individual cluster fastas, not meant for external use
}
\keyword{clustering}
