% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{runBlast}
\alias{runBlast}
\title{Function to blast significant sequences}
\usage{
runBlast(blast_exec, blast_db, ncores, sigResults, map)
}
\description{
Function to blast significant sequences
}
\examples{
runBlast(...)
}
\keyword{filtering}
