% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{runBlast}
\alias{runBlast}
\title{Function to blast small RNA sequences.}
\usage{
runBlast(blast_exec, blast_db, ncores, sig_sequences, strand = "plus")
}
\arguments{
\item{blast_exec}{path to your installation of 'blastn'}

\item{blast_db}{path to your local blastdb}

\item{ncores}{number of cores used for blasting}

\item{sig_sequences}{vector of sequences in fasta style generated by the sequencesAsFasta}

\item{strand}{query strand(s) to search against database, default = 'plus'}
}
\description{
This function takes a vector of sequences in fasta style as input and blasts them using a local blast database.
The parameters are configured for blasting small RNA sequences with 100\% identity over the whole query sequence.
}
\examples{
blast_exec <- "$HOME/software/ncbi-blast-2.6.0+/bin/blastn"
blast_db <-"$HOME/blastdb/MouseDB.fa"
ncores <- 2
testseq <- c(">seq_1","GCATTGGTGGTTCAGTGGTAGAATTCTCGCC")
blastResult <- runBlast(blast_exec, blast_db, ncores, testseq)
}
\keyword{blastn}
