% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{runDESeq2}
\alias{runDESeq2}
\title{Function to run differential expression analysis based on DESeq2}
\usage{
runDESeq2(countData, phenoData, design, map, out_dir)
}
\arguments{
\item{countData}{raw counts of sequences}

\item{phenoData}{data frame with sample names as rownames and assigned condition in first column}

\item{design}{design formular for DE analysis}

\item{map}{data frame with sequences as rownames and sequence ID in first column}

\item{out_dir}{directory for sample distance map and MA plot}
}
\description{
Function to run differential expression analysis based on DESeq2
}
\examples{
in_dir <- system.file("extdata", package = "DEUS")
out_dir <- "$HOME/out"
phenofile <- system.file("extdata", "condition_test.tsv", package = "DEUS")
phenoInfo <- read.table(phenofile, header=T, row.names=1, check.names=FALSE)
countTable <- createCountTableFromFastQs(in_dir, phenoInfo=phenoInfo)
countTable <- filterLowExp(countTable, phenoInfo)
design <- ~ condition
deResults <- runDESeq2(countTable, phenoInfo, design, map, out_dir)
}
\keyword{analysis}
\keyword{differential}
\keyword{expression}
