% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{runDESeq2}
\alias{runDESeq2}
\title{Run differential expression analysis based on DESeq2}
\usage{
runDESeq2(count_table, pheno_info, design, contrast, effect_name,
  shrink_type = "normal", out_dir, prefix = "DESeq2")
}
\arguments{
\item{count_table}{The count table generated by \link[DEUS]{createCountTableFromFastQs}}

\item{pheno_info}{A data frame with sample identifiers as row names including the columns used for the design formula.
The sample identifiers must be identical to those in the count_table.}

\item{design}{Design formula for differential expression analysis}

\item{contrast}{Argument to specify the comparison for which the results table will be generated (corresponds to 'contrast' argument in \link[DESeq2]{results} and \link[DESeq2]{lfcShrink}). Required if interaction terms are part of the design formula.}

\item{effect_name}{Argument to specify the comparision for which the results table will be generated (corresponds to 'name' argument in \link[DESeq2]{results} and 'coef' argument in \link[DESeq2]{lfcShrink}). Required if interaction terms are part of the design formula.}

\item{shrink_type}{DESeq2 shrinkage estimator (corresponds to 'type' argument in \link[DESeq2]{lfcShrink}).}

\item{out_dir}{Directory to save sample distance map, PCA and MA plot.}

\item{prefix}{Prefix for the names of the PDF files containing the sample distance map, the PCA and the MA plot.}
}
\value{
Returns a list consisting of 'normCounts' and the 'deResult'.
'normCounts' is a data frame of normalized counts with sequences as row names.
'deResult' is a data frame with sequences as row names and the columns 'pvalue', 'padj', 'baseMean', 'log2FoldChange', 'lfcSE', 'stat', 'IHWPval'.
See \link[DESeq2]{results} and \link[IHW]{ihw.default} for further documentation on the columns.
}
\description{
The function provides a reference implementation of differential expression analysis using \link[DESeq2]{DESeq}.
It is possible to replace this function by any other method generating at least the log2 fold change, a p-value and an IHW p-value for each unique sequence to be compatible with downstream functions.
}
