% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{runDESeq2}
\alias{runDESeq2}
\title{Run differential expression analysis based on DESeq2}
\usage{
runDESeq2(count_table, pheno_info, design, out_dir = NULL)
}
\arguments{
\item{count_table}{The count table generated by \link[DEUS]{createCountTableFromFastQs}}

\item{pheno_info}{A data frame with sample identifiers as row names including the columns used for the design formula.
The sample identifiers must be identical to those in the count_table.}

\item{design}{Design formula for differential expression analysis}

\item{out_dir}{Directory to save sample distance map, PCA and MA plot.}
}
\value{
Returns a list consisting of 'normCounts' and the 'deResult'.
'normCounts' is a data frame of normalized counts with sequences as row names.
'deResult' is a data frame with sequences as row names and the columns 'pvalue', 'padj', 'baseMean', 'log2FoldChange', 'lfcSE', 'stat', 'IHWPval'.
See \link[DESeq2]{results} and \link[IHW]{ihw} for further documentation on the columns.
}
\description{
The function provides a reference implementation of differential expression analysis using \link[DESeq2]{DESeq}.
It is possible to replace this function by any other method generating at least the log2 fold change, a p-value and an IHW p-value for each unique sequence to be compatible with downstream functions.
}
