% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sequencesAsFasta}
\alias{sequencesAsFasta}
\title{Function to get sequences as a vector of fasta IDs and sequences}
\usage{
sequencesAsFasta(sigResults, map)
}
\arguments{
\item{sigResults}{resulting table of differential expression analysis with significant sequences as row names}

\item{map}{map of sequences to sequence IDs used as IDs in fasta}
}
\description{
Function to get sequences as a vector of fasta IDs and sequences
}
\keyword{fasta}
