% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{writeSummaryFiles}
\alias{writeSummaryFiles}
\title{Function to write summary tables and fasta files for USBseq result}
\usage{
writeSummaryFiles(summaryTable, outDir)
}
\arguments{
\item{summaryTable}{summary table generated by mergeResults function (requires blast result)}

\item{outDir}{directory for summary files (tsv tables and fasta files)}
}
\description{
Function to write summary tables and fasta files for USBseq result
}
\examples{

}
\keyword{summary}
